git-svn-id: file:///home/danlipsa/dan-svn-repo/src@1437 d596a3c1-de11-4375-84b4-10211c464620

+++ Histogram.h
@@ -41,11 +41,11 @@ public:
     {
 	m_histogramItem.getSelectedBins (intervals, selected);
     }
-    size_t GetAxisMaxValue () const
+    double GetAxisMaxValue () const
     {
 	return m_axisMaxValue;
     }
-    void SetAxisMaxValue (size_t axisMaxValue);
+    void SetAxisMaxValue (double axisMaxValue);
     size_t GetDataMaxValue () const;
 
 public Q_SLOTS:
@@ -68,7 +68,7 @@ private:
     QwtPlotPicker m_plotPicker;
     size_t m_beginBinSelection;
     SelectionTool m_selectionTool;
-    size_t m_axisMaxValue;
+    double m_axisMaxValue;
 };
